backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-genai
pytest
pytest-asyncio
httpx
python-dotenv
psycopg2-binary
